Update LookML tests
